Separate worker methods from server controller
